+++ src/ctorch.cpp
@@ -13,6 +13,25 @@
 #endif
 
 // =============================================================================
+// --- Functions to aid in consistent error handling
+// =============================================================================
+
+// Accept a string message and handle as error. Accepts a cleanup function if desired.
+void ctorch_error(const std::string &message,
+                  const std::function<void()> &cleanup = nullptr) {
+  std::cerr << "[ERROR]: " << message << std::endl;
+  if (cleanup) {
+    cleanup(); // Perform cleanup actions
+  }
+  exit(EXIT_FAILURE);
+}
+
+// Accept a string message and handle as a warning.
+void ctorch_warn(const std::string &message) {
+  std::cerr << "[WARNING]: " << message << std::endl;
+}
+
+// =============================================================================
 // --- Constant expressions
 // =============================================================================
 
@@ -20,9 +39,8 @@
 constexpr auto get_libtorch_dtype(torch_data_t dtype) {
   switch (dtype) {
   case torch_kUInt8:
-    std::cerr << "[WARNING]: uint8 not supported in Fortran" << std::endl;
+    ctorch_error("uint8 not supported in Fortran");
     // See https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
-    exit(EXIT_FAILURE);
   case torch_kInt8:
     return torch::kInt8;
   case torch_kInt16:
@@ -32,15 +50,14 @@ constexpr auto get_libtorch_dtype(torch_data_t dtype) {
   case torch_kInt64:
     return torch::kInt64;
   case torch_kFloat16:
-    std::cerr << "[WARNING]: float16 not supported in Fortran" << std::endl;
+    ctorch_error("float16 not supported in Fortran");
     // See https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
-    exit(EXIT_FAILURE);
   case torch_kFloat32:
     return torch::kFloat32;
   case torch_kFloat64:
     return torch::kFloat64;
   default:
-    std::cerr << "[WARNING]: unknown data type, setting to torch_kFloat32" << std::endl;
+    ctorch_warn("unknown data type, setting to torch_kFloat32");
     return torch::kFloat32;
   }
 }
@@ -48,9 +65,8 @@ constexpr auto get_libtorch_dtype(torch_data_t dtype) {
 // Mapping from libtorch Dtype to FTorch device_data_t
 torch_data_t get_ftorch_dtype(caffe2::TypeMeta dtype) {
   if (dtype == torch::kUInt8) {
-    std::cerr << "[WARNING]: uint8 not supported in Fortran" << std::endl;
+    ctorch_error("uint8 not supported in Fortran");
     // See https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
-    exit(EXIT_FAILURE);
   } else if (dtype == torch::kInt8) {
     return torch_kInt8;
   } else if (dtype == torch::kInt16) {
@@ -60,9 +76,8 @@ torch_data_t get_ftorch_dtype(caffe2::TypeMeta dtype) {
   } else if (dtype == torch::kInt64) {
     return torch_kInt64;
   } else if (dtype == torch::kFloat16) {
-    std::cerr << "[WARNING]: float16 not supported in Fortran" << std::endl;
+    ctorch_error("float16 not supported in Fortran");
     // See https://gcc.gnu.org/onlinedocs/gfortran/ISO_005fFORTRAN_005fENV.html
-    exit(EXIT_FAILURE);
   } else if (dtype == torch::kFloat32) {
     return torch_kFloat32;
   } else if (dtype == torch::kFloat64) {
@@ -72,6 +87,7 @@ torch_data_t get_ftorch_dtype(caffe2::TypeMeta dtype) {
               << std::endl;
     exit(EXIT_FAILURE);
   }
+  return torch_kFloat32; // Dummy return to satisfy the compiler
 }
 
 // Mapping from FTorch device_type_t to libtorch DeviceType
@@ -79,13 +95,13 @@ const auto get_libtorch_device(torch_device_t device_type, int device_index) {
   switch (device_type) {
   case torch_kCPU:
     if (device_index != -1) {
-      std::cerr << "[WARNING]: device index unused for CPU-only runs" << std::endl;
+      ctorch_warn("device index unused for CPU-only runs");
     }
     return torch::Device(torch::kCPU);
 #if GPU_DEVICE == GPU_DEVICE_CUDA
   case torch_kCUDA:
     if (device_index == -1) {
-      std::cerr << "[WARNING]: device index unset, defaulting to 0" << std::endl;
+      ctorch_warn("device index unset, defaulting to 0");
       device_index = 0;
     }
     if (device_index >= 0 && device_index < torch::cuda::device_count()) {
@@ -98,13 +114,13 @@ const auto get_libtorch_device(torch_device_t device_type, int device_index) {
 #endif
   case torch_kMPS:
     if (device_index != -1 && device_index != 0) {
-      std::cerr << "[WARNING]: Only one device is available for MPS runs" << std::endl;
+      ctorch_warn("Only one device is available for MPS runs");
     }
     return torch::Device(torch::kMPS);
 #if GPU_DEVICE == GPU_DEVICE_XPU
   case torch_kXPU:
     if (device_index == -1) {
-      std::cerr << "[WARNING]: device index unset, defaulting to 0" << std::endl;
+      ctorch_warn("device index unset, defaulting to 0");
       device_index = 0;
     }
     if (device_index >= 0 && device_index < torch::xpu::device_count()) {
@@ -116,7 +132,7 @@ const auto get_libtorch_device(torch_device_t device_type, int device_index) {
     }
 #endif
   default:
-    std::cerr << "[WARNING]: unknown device type, setting to torch_kCPU" << std::endl;
+    ctorch_warn("unknown device type, setting to torch_kCPU");
     return torch::Device(torch::kCPU);
   }
 }
@@ -140,6 +156,43 @@ const torch_device_t get_ftorch_device(torch::DeviceType device_type) {
 }
 
 // =============================================================================
+// --- Functions for validating tensors
+// =============================================================================
+
+// Check if a tensor is valid
+void validate_tensor_not_null(const torch::Tensor *t, const std::string &name) {
+  if (!t) {
+    throw std::invalid_argument(name + " is null.");
+  }
+}
+
+// Check if a tensor is defined
+void validate_tensor_defined(const torch::Tensor *t, const std::string &name) {
+  if (!t->defined()) {
+    throw std::invalid_argument(name + " is undefined.");
+  }
+}
+
+void validate_tensor(const torch::Tensor *t, const std::string &name) {
+  validate_tensor_not_null(t, name);
+  validate_tensor_defined(t, name);
+}
+
+// Check if a tensor has requires_grad set
+void validate_requires_grad(const torch::Tensor *t, const std::string &name) {
+  if (!t->requires_grad()) {
+    throw std::runtime_error(name + " does not have requires_grad set.");
+  }
+}
+
+void validate_gradient_defined(const torch::Tensor *t, const std::string &name) {
+  if (!t->grad().defined()) {
+    throw std::runtime_error(
+        name + " has an undefined gradient.\nPerhaps you forgot to call backward.");
+  }
+}
+
+// =============================================================================
 // --- Functions for constructing tensors
 // =============================================================================
 
@@ -147,7 +200,7 @@ torch_tensor_t torch_empty(int ndim, const int64_t *shape, torch_data_t dtype,
                            torch_device_t device_type, int device_index = -1,
                            const bool requires_grad = false) {
   torch::AutoGradMode enable_grad(requires_grad);
-  torch::Tensor *tensor = nullptr;
+  auto tensor = new torch::Tensor;
   try {
     // This doesn't throw if shape and dimensions are incompatible
     c10::IntArrayRef vshape(shape, ndim);
@@ -155,16 +208,11 @@ torch_tensor_t torch_empty(int ndim, const int64_t *shape, torch_data_t dtype,
                        .dtype(get_libtorch_dtype(dtype))
                        .device(get_libtorch_device(device_type, device_index))
                        .requires_grad(requires_grad);
-    tensor = new torch::Tensor;
     *tensor = torch::empty(vshape, options);
   } catch (const torch::Error &e) {
-    std::cerr << "[ERROR]: " << e.msg() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.msg(), [&]() { delete tensor; });
   } catch (const std::exception &e) {
-    std::cerr << "[ERROR]: " << e.what() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.what(), [&]() { delete tensor; });
   }
   return tensor;
 }
@@ -173,7 +221,7 @@ torch_tensor_t torch_zeros(int ndim, const int64_t *shape, torch_data_t dtype,
                            torch_device_t device_type, int device_index = -1,
                            const bool requires_grad = false) {
   torch::AutoGradMode enable_grad(requires_grad);
-  torch::Tensor *tensor = nullptr;
+  auto tensor = new torch::Tensor;
   try {
     // This doesn't throw if shape and dimensions are incompatible
     c10::IntArrayRef vshape(shape, ndim);
@@ -181,16 +229,11 @@ torch_tensor_t torch_zeros(int ndim, const int64_t *shape, torch_data_t dtype,
                        .dtype(get_libtorch_dtype(dtype))
                        .device(get_libtorch_device(device_type, device_index))
                        .requires_grad(requires_grad);
-    tensor = new torch::Tensor;
     *tensor = torch::zeros(vshape, options);
   } catch (const torch::Error &e) {
-    std::cerr << "[ERROR]: " << e.msg() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.msg(), [&]() { delete tensor; });
   } catch (const std::exception &e) {
-    std::cerr << "[ERROR]: " << e.what() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.what(), [&]() { delete tensor; });
   }
   return tensor;
 }
@@ -199,7 +242,7 @@ torch_tensor_t torch_ones(int ndim, const int64_t *shape, torch_data_t dtype,
                           torch_device_t device_type, int device_index = -1,
                           const bool requires_grad = false) {
   torch::AutoGradMode enable_grad(requires_grad);
-  torch::Tensor *tensor = nullptr;
+  auto tensor = new torch::Tensor;
   try {
     // This doesn't throw if shape and dimensions are incompatible
     c10::IntArrayRef vshape(shape, ndim);
@@ -207,16 +250,11 @@ torch_tensor_t torch_ones(int ndim, const int64_t *shape, torch_data_t dtype,
                        .dtype(get_libtorch_dtype(dtype))
                        .device(get_libtorch_device(device_type, device_index))
                        .requires_grad(requires_grad);
-    tensor = new torch::Tensor;
     *tensor = torch::ones(vshape, options);
   } catch (const torch::Error &e) {
-    std::cerr << "[ERROR]: " << e.msg() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.msg(), [&]() { delete tensor; });
   } catch (const std::exception &e) {
-    std::cerr << "[ERROR]: " << e.what() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.what(), [&]() { delete tensor; });
   }
   return tensor;
 }
@@ -228,7 +266,7 @@ torch_tensor_t torch_from_blob(void *data, int ndim, const int64_t *shape,
                                torch_device_t device_type, int device_index = -1,
                                const bool requires_grad = false) {
   torch::AutoGradMode enable_grad(requires_grad);
-  torch::Tensor *tensor = nullptr;
+  auto tensor = new torch::Tensor;
 
   try {
     // This doesn't throw if shape and dimensions are incompatible
@@ -238,17 +276,12 @@ torch_tensor_t torch_from_blob(void *data, int ndim, const int64_t *shape,
                        .dtype(get_libtorch_dtype(dtype))
                        .device(get_libtorch_device(device_type, device_index))
                        .requires_grad(requires_grad);
-    tensor = new torch::Tensor;
     *tensor = torch::from_blob(data, vshape, vstrides, options);
 
   } catch (const torch::Error &e) {
-    std::cerr << "[ERROR]: " << e.msg() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.msg(), [&]() { delete tensor; });
   } catch (const std::exception &e) {
-    std::cerr << "[ERROR]: " << e.what() << std::endl;
-    delete tensor;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.what(), [&]() { delete tensor; });
   }
   return tensor;
 }
@@ -314,6 +347,7 @@ void torch_tensor_delete(torch_tensor_t tensor) {
 
 void torch_tensor_zero(torch_tensor_t tensor) {
   auto t = reinterpret_cast<torch::Tensor *>(tensor);
+  validate_tensor(t, "Input tensor");
   t->zero_();
 }
 
@@ -324,6 +358,8 @@ void torch_tensor_zero(torch_tensor_t tensor) {
 void torch_tensor_assign(torch_tensor_t output, const torch_tensor_t input) {
   auto out = reinterpret_cast<torch::Tensor *>(output);
   auto in = reinterpret_cast<torch::Tensor *const>(input);
+  validate_tensor(out, "Output tensor");
+  validate_tensor(in, "Input tensor");
   torch::AutoGradMode enable_grad(in->requires_grad());
   // NOTE: The following line ensures that the output tensor continues to point to a
   //       Fortran array if it was set up to do so using torch_tensor_from_array. If
@@ -388,12 +424,55 @@ void torch_tensor_power_float(torch_tensor_t output, const torch_tensor_t tensor
   *out = pow(*t, *exp);
 }
 
+// ============================================================================
+// --- Other operators for computations involving tensors
+// ============================================================================
+
+void torch_tensor_sum(torch_tensor_t output, const torch_tensor_t tensor) {
+  auto out = reinterpret_cast<torch::Tensor *>(output);
+  auto t = reinterpret_cast<torch::Tensor *const>(tensor);
+
+  if (torch_tensor_get_rank(output) != 1) {
+    std::stringstream errmsg;
+    errmsg << "Invalid rank of output tensor for sum\nrank="
+           << torch_tensor_get_rank(output) << " != 1";
+    ctorch_error(errmsg.str());
+  }
+  if (torch_tensor_get_sizes(output)[0] != 1) {
+    std::stringstream errmsg;
+    errmsg << "Invalid shape of output tensor for sum\nshape=["
+           << torch_tensor_get_sizes(output)[0] << "] != [1]";
+    ctorch_error(errmsg.str());
+  }
+  std::move(*out) = t->sum();
+}
+
+void torch_tensor_mean(torch_tensor_t output, const torch_tensor_t tensor) {
+  auto out = reinterpret_cast<torch::Tensor *>(output);
+  auto t = reinterpret_cast<torch::Tensor *const>(tensor);
+
+  if (torch_tensor_get_rank(output) != 1) {
+    std::stringstream errmsg;
+    std::cerr << "Invalid rank of output tensor for mean\nrank="
+              << torch_tensor_get_rank(output) << " != 1";
+    ctorch_error(errmsg.str());
+  }
+  if (torch_tensor_get_sizes(output)[0] != 1) {
+    std::stringstream errmsg;
+    errmsg << "Invalid shape of output tensor for mean\nshape=["
+           << torch_tensor_get_sizes(output)[0] << "] != [1]";
+    ctorch_error(errmsg.str());
+  }
+  std::move(*out) = t->mean();
+}
+
 // =============================================================================
 // --- Functions related to automatic differentiation functionality for tensors
 // =============================================================================
 
 void torch_tensor_zero_grad(torch_tensor_t tensor) {
   auto t = reinterpret_cast<torch::Tensor *>(tensor);
+  validate_tensor(t, "Gradient to zero");
   t->mutable_grad().zero_();
 }
 
@@ -402,13 +481,37 @@ void torch_tensor_backward(const torch_tensor_t tensor,
                            const bool retain_graph) {
   auto t = reinterpret_cast<torch::Tensor *>(tensor);
   auto g = reinterpret_cast<torch::Tensor *const>(external_gradient);
-  t->backward(*g, retain_graph);
+
+  try {
+    // Check if the tensors are valid and defined
+    validate_tensor(t, "Input tensor");
+    validate_tensor(g, "External gradient");
+
+    // Perform backwards step
+    t->backward(*g, retain_graph);
+  } catch (const std::exception &e) {
+    ctorch_error(std::string(e.what()) + " in torch_tensor_backward");
+  }
 }
 
 void torch_tensor_get_gradient(const torch_tensor_t tensor, torch_tensor_t gradient) {
-  auto t = reinterpret_cast<torch::Tensor *const>(tensor);
-  auto g = reinterpret_cast<torch::Tensor *>(gradient);
-  std::move(*g) = t->grad();
+  try {
+    // Cast the input pointers to torch::Tensor
+    auto t = reinterpret_cast<torch::Tensor *const>(tensor);
+    auto g = reinterpret_cast<torch::Tensor *>(gradient);
+
+    // Check if the tensors are valid and defined
+    validate_tensor(t, "Input tensor");
+    validate_tensor_not_null(g, "Output gradient");
+    // Check input has requires_grad set and can generate a valid gradient tensor
+    validate_requires_grad(t, "Input tensor");
+    validate_gradient_defined(t, "Input tensor");
+
+    // Assign the gradient to the output tensor
+    std::move(*g) = t->grad();
+  } catch (const std::exception &e) {
+    ctorch_error(std::string(e.what()) + " in torch_tensor_get_gradient");
+  }
 }
 
 // =============================================================================
@@ -436,13 +539,9 @@ torch_jit_script_module_t torch_jit_load(const char *filename,
     *module =
         torch::jit::load(filename, get_libtorch_device(device_type, device_index));
   } catch (const torch::Error &e) {
-    std::cerr << "[ERROR]: " << e.msg() << std::endl;
-    delete module;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.msg(), [&]() { delete module; });
   } catch (const std::exception &e) {
-    std::cerr << "[ERROR]: " << e.what() << std::endl;
-    delete module;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.what(), [&]() { delete module; });
   }
   set_is_training(module, is_training);
 
@@ -469,10 +568,7 @@ void torch_jit_module_forward(const torch_jit_script_module_t module,
     if (LocalTensor.isTensor()) {
       inputs_vec.push_back(LocalTensor);
     } else {
-      std::cerr << "[ERROR]: One of the inputs to torch_jit_module_forward is "
-                   "not a Tensor."
-                << std::endl;
-      exit(EXIT_FAILURE);
+      ctorch_error("One of the inputs to torch_jit_module_forward is not a Tensor");
     }
   }
   try {
@@ -488,14 +584,12 @@ void torch_jit_module_forward(const torch_jit_script_module_t module,
     } else {
       // If for some reason the forward method does not return a Tensor it
       // should raise an error when trying to cast to a Tensor type
-      std::cerr << "[ERROR]: Model Output is neither Tensor nor Tuple." << std::endl;
+      ctorch_error("Model Output is neither Tensor nor Tuple");
     }
   } catch (const torch::Error &e) {
-    std::cerr << "[ERROR]: " << e.msg() << std::endl;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.msg());
   } catch (const std::exception &e) {
-    std::cerr << "[ERROR]: " << e.what() << std::endl;
-    exit(EXIT_FAILURE);
+    ctorch_error(e.what());
   }
 }